abstracted doesPrereqExist
